Cleaning JS
