+++ src/components/App.js
@@ -10,10 +10,12 @@ function App() {
       if (user) {
         setUserObj({
           displayName: user.displayName,
-          uid: userObj.uid,
+          uid: user.uid,
           updateProfile: (args) => user.updateProfile(args),
         });
-      } 
+      } else {
+        setUserObj(null);
+      }
       setInit(true);
     });
   }, []);
@@ -21,7 +23,7 @@ function App() {
     const user = authService.currentUser;
     setUserObj({
       displayName: user.displayName,
-      uid: userObj.uid,
+      uid: user.uid,
       updateProfile: (args) => user.updateProfile(args),
     });
   }
