+++ src/routes/Auth.js
@@ -1,69 +1,30 @@
 import React, { useState } from "react";
 import { authService, firebaseInstance } from "fbase";
+import AuthForm from "components/AuthForm";
 
 const Auth = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [newAccount, setNewAccount] = useState(true);
-  const [error, setError] = useState("");
-  const onChange = (event) => {
-    const {target: {name, value}} = event;
-    if(name === "email"){
-      setEmail(value);
-    } else if (name === "password"){
-      setPassword(value);
-    }
-  };
-  const onSubmit = async (event) => {
-    event.preventDefault();
-    let data;
-    try{
-      if(newAccount){
-        data = await authService.createUserWithEmailAndPassword(
-          email, 
-          password
-        );
-      } else {
-        data = await authService.signInWithEmailAndPassword(
-          email, 
-          password
-        );
-      }
-      console.log(data);
-    } catch(error){
-      setError(error.message);
-    }
-  }
-  const toggleAccount = () => setNewAccount((prev) => !prev);
-  const onSocialClick = async (event) => {
+  const onSocialClick = async event => {
     const {
-      target:{
-        name
-      }
+      target: { name }
     } = event;
     let provider;
-    if(name === "google"){
+    if (name === "google") {
       provider = new firebaseInstance.auth.GoogleAuthProvider();
-    }else if(name === "github"){
+    } else if (name === "github") {
       provider = new firebaseInstance.auth.GithubAuthProvider();
     }
-    const data = await authService.signInWithPopup(provider);
-    console.log(data);
-  }
+    await authService.signInWithPopup(provider);
+  };
   return (
     <div>
-      <form onSubmit={onSubmit}>
-        <input name="email" type="email" placeholder="Email" required value={email} onChange={onChange} />
-        <input name="password" type="password" placeholder="Password" required value={password} onChange={onChange} />
-        <input type="submit" value={newAccount ? "Create Account" : "Log In"} />
-        {error}
-      </form>
-      <span onClick={toggleAccount}>
-        {newAccount ? "Sign in" : "Create Account"}
-        </span>
+      <AuthForm />
       <div>
-        <button onClick={onSocialClick} name="google">Continue with Google</button>
-        <button onClick={onSocialClick} name="github">Continue with Github</button>
+        <button onClick={onSocialClick} name="google">
+          Continue with Google
+        </button>
+        <button onClick={onSocialClick} name="github">
+          Continue with Github
+        </button>
       </div>
     </div>
   );
