+++ src/routes/Home.js
@@ -1,12 +1,10 @@
 import Nweet from "components/Nweet";
+import NweetFactory from "components/NweetFactory";
 import { dbService, storageService } from "fbase";
 import React, { useEffect, useState } from "react";
-import { v4 as uuidV4 } from "uuid";
 
 const Home = ({ userObj }) => {
-  const [nweet, setNweet] = useState("");
   const [nweets, setNweets] = useState([]);
-  const [attachment, setAttachment] = useState("");
   useEffect(() => {
     dbService.collection("nweets").onSnapshot(snapshot => {
       const nweetArray = snapshot.docs.map(doc => ({
@@ -16,66 +14,9 @@ const Home = ({ userObj }) => {
       setNweets(nweetArray);
     });
   }, []);
-  const onSubmit = async event => {
-    event.preventDefault();
-    let attachmentUrl = "";
-    if(attachment !== ""){
-      const attachementRef = storageService
-        .ref()
-        .child(`${userObj.uid}/${uuidV4()}`);
-      const response = await attachementRef.putString(attachment, "data_url");
-      attachmentUrl = await response.ref.getDownloadURL();
-    } 
-    const nweetObj = {
-      text: nweet,
-      createAt: Date.now(),  
-      creatorId: userObj.uid,
-      attachmentUrl
-    };
-    await dbService.collection("nweets").add(nweetObj);
-    setNweet("");
-    setAttachment("");
-  };
-  const onChange = event => {
-    const {
-      target: { value }
-    } = event;
-    setNweet(value);
-  };
-  const onFileChange = (event) =>{
-    const {
-      target: { files }
-    } = event;
-    const theFile = files[0];
-    const reader = new FileReader();
-    reader.onloadend = (finishedEvent) => {
-      const {
-        currentTarget: { result }
-      } = finishedEvent;
-      setAttachment(result)
-    }
-    reader.readAsDataURL(theFile);
-  }
-  const onClearAttachment = () => setAttachment(null);
   return (
     <div>
-      <form onSubmit={onSubmit}>
-        <input
-          value={nweet}
-          onChange={onChange}
-          type="text"
-          placeholder="What's on your mind?"
-          maxLength={120}
-        />
-        <input type="file" accept="image/*" onChange={onFileChange} />
-        <input type="submit" value="Nweet" />
-        {attachment && (
-          <div>
-            <img src={attachment} width="50px" height="50px" />
-            <button onClick={onClearAttachment}>Clear</button>
-          </div>
-        )}
-      </form>
+      <NweetFactory userObj={userObj}/>
       <div>
         {nweets.map(nweet => (
           <Nweet
